Move components out of grouped and nongrouped
